+++ packages/transaction-pay-controller/src/strategy/fiat/FiatStrategy.test.ts
@@ -3,7 +3,7 @@ import type { TransactionMeta } from '@metamask/transaction-controller';
 import { getFiatQuotes } from './fiat-quotes';
 import { submitFiatQuotes } from './fiat-submit';
 import { FiatStrategy } from './FiatStrategy';
-import type { FiatOriginalQuote } from './types';
+import type { FiatQuote } from './types';
 import type { TransactionPayControllerMessenger } from '../..';
 import type { TransactionPayQuote } from '../../types';
 
@@ -12,7 +12,7 @@ jest.mock('./fiat-submit');
 
 const QUOTE_MOCK = {
   estimatedDuration: 5,
-} as TransactionPayQuote<FiatOriginalQuote>;
+} as TransactionPayQuote<FiatQuote>;
 
 describe('FiatStrategy', () => {
   const getFiatQuotesMock = jest.mocked(getFiatQuotes);