+++ packages/transaction-pay-controller/src/strategy/fiat/types.ts
@@ -1,8 +1,10 @@
-import type { Quote } from '@metamask/ramps-controller';
+import type { Quote, QuotesResponse } from '@metamask/ramps-controller';
 
 import type { RelayQuote } from '../relay/types';
 
-export type FiatOriginalQuote = {
-  fiatQuote: Quote;
+export type FiatQuote = {
+  rampsQuote: Quote;
   relayQuote: RelayQuote;
 };
+
+export type FiatQuotesResponse = QuotesResponse;